+++ lib/character.dart
@@ -1,3 +1,5 @@
+import 'package:parser_builder/transformers.dart';
+
 import 'builder_helper.dart' as helper;
 import 'parser_builder.dart';