+++ lib/src/character/alphanumeric1.dart
@@ -12,7 +12,6 @@ class Alphanumeric1 extends _Chars1 {
 
   @override
   Transformer<int, bool> _getCharacterPredicate() {
-    return Transformer('c',
-        '=> c >= 0x30 && c <= 0x39 || c >= 0x41 && c <= 0x5a || c >= 0x61 && c <= 0x7a;');
+    return CharClass('[#x30-#x39] | [#x41-#x5a] | [#x61-#x7a]');
   }
 }